titanmanportal/requirements/requirements-DEV.txt
django==3.0.4
django-grappelli==2.14.1
django-phonenumber-field==4.0.0
phonenumbers==8.12.0
pillow==7.0
python-dotenv==0.12.0
pyyaml==5.3.1